==> models/notebooks/02_geospatial_analysis.ipynb <==
import pandas as pd
import os
from shapely.geometry import Point, Polygon
from shapely import geometry
import ast

#### Reading In Relevant Files

directory = '../../data/processed'
void = ['clean_areas.csv', 'clean_disadvantaged_areas.csv', 'clean_police_districts.csv', 'clean_public_healthindicator.csv', 'clean_train_ridership.csv', 'clean_bike_trips.csv']

# read in the code files
def read_in(directory, void):
    data = {}
    for filename in os.listdir(directory):
        if filename not in void and filename.endswith('.csv'):
            file_path = os.path.join(directory, filename)
            data[filename[:-4]] = pd.read_csv(file_path)
            print(f'{filename[:-4]} successfully read in')
    return data

data = read_in(directory, void)

clean_police_districts = pd.read_csv('../../data/processed/clean_police_districts.csv')
clean_areas = pd.read_csv('../../data/processed/clean_areas.csv')
clean_disadvantaged_areas = pd.read_csv('../../data/processed/clean_disadvantaged_areas.csv')

#### Data Cleaning of Geom Types

def parse_polygon1(polygon_string):
    points = polygon_string.strip('POLYGON ((').strip('))').split(', ')
    points = [tuple(map(float, point.split())) for point in points]
    return Polygon(points)

def parse_polygon2(polygon_string):
    points = ast.literal_eval(polygon_string)
    return Polygon(points)

def swap_coordinates(polygon):
    if polygon.is_empty:
        return polygon
    swapped_coords = [(y, x) for x, y in polygon.exterior.coords]
    return Polygon(swapped_coords)

clean_police_districts['geom'] = clean_police_districts['geom'].apply(parse_polygon1)
clean_police_districts.head()

clean_areas['poly'] = clean_areas['poly'].apply(parse_polygon2)
clean_areas.head()

clean_disadvantaged_areas['poly'] = clean_disadvantaged_areas['poly'].apply(parse_polygon1)
clean_disadvantaged_areas['poly'] = clean_disadvantaged_areas['poly'].apply(swap_coordinates)
clean_disadvantaged_areas.head()

clean_areas['centroid'] = clean_areas['poly'].apply(lambda poly : poly.centroid)
clean_areas

clean_disadvantaged_areas['centroid'] = clean_disadvantaged_areas['poly'].apply(lambda poly : poly.centroid)
clean_disadvantaged_areas

#### Assigning Districts to Each Dataset

def determine_within(df):

    statuses = []
    districts = []
    cent = True if 'centroid' in df.columns else False

    for i in range(len(df)):
        point = df.centroid.loc[i] if cent else geometry.Point(df.long.loc[i], df.lat.loc[i])
        status = 0

        for index, row in clean_police_districts.iterrows():
            district = row['district']
            geom = row['geom']
            if (geom.contains(point)): 
                status = 1
                break
        statuses.append(status)
        districts.append(district)

    df['status'] = statuses
    df['district'] = districts
    df = df[df['status'] == 1].drop('status', axis=1)

    if cent: df.drop('centroid', axis=1, inplace=True)

    return df

def determine_districts(data):
    for df_name, df in data.items():
        data[df_name] = determine_within(df)
        print(f'{df_name} successfully completed')
    return data

clean_disadvantaged_areas = determine_within(clean_disadvantaged_areas)
clean_areas = determine_within(clean_areas)

clean_areas

#data = determine_districts(data)

#### Assigning Areas Stats to Each Crime

def determine_area_for_crimes(df):

    statuses = []
    areas = []
    cent = True if 'centroid' in df.columns else False
    for i in range(len(df)):
        point = df.loc[i, 'centroid'] if cent else geometry.Point(df.loc[i, 'long'], df.loc[i, 'lat'])
        status = 0

        for index, row in clean_areas.iterrows():
            area = row['id']
            geom = row['poly']
            if (geom.contains(point)): 
                status = 1
                break
        statuses.append(status)
        areas.append(area)

    df['status'] = statuses
    df['areas'] = areas
    df = df[df['status'] == 1].drop('status', axis=1)

    if cent: df.drop('centroid', axis=1, inplace=True)

    return df

#data['clean_crime'] = determine_area_for_crimes(data['clean_crime'])

clean_bike_trips = pd.read_csv('../../data/processed/clean_bike_trips.csv')
clean_train_ridership = pd.read_csv('../../data/processed/clean_train_ridership.csv')

clean_disadvantaged_areas['centroid'] = clean_disadvantaged_areas['poly'].apply(lambda poly : poly.centroid)
clean_disadvantaged_areas

clean_disadvantaged_areas['poly'].drop_duplicates(inplace=True)
clean_disadvantaged_areas.reset_index(drop=True, inplace=True)

trains_with_areas = determine_area_for_crimes(data['clean_train_stations'])
bikes_with_areas = determine_area_for_crimes(data['clean_bike_stations'])
disadvantaged_areas_within_areas = determine_area_for_crimes(clean_disadvantaged_areas)

#### Data Cleaning to Finalize Datasets

clean_bike_trips = clean_bike_trips[['station_id','date','lat','long']].merge(right=bikes_with_areas[['id', 'district', 'areas']], how='left', left_on='station_id', right_on='id').dropna(subset=['district'])
clean_train_ridership = clean_train_ridership[['date','line','station_name','lat','long', 'rides']].merge(right=trains_with_areas[['station_name', 'district', 'areas']], how='left', on='station_name').dropna(subset=['district'])

clean_police_districts = clean_police_districts.merge(right=clean_disadvantaged_areas.groupby('district').agg('count'), how='left', on='district').fillna(0).rename(columns={'poly':'disadvantaged_score'})
clean_police_districts

clean_public_healthindicator = pd.read_csv('../../data/processed/clean_public_healthindicator.csv')

agg_public_healthindicator = pd.merge(left=clean_public_healthindicator, right=clean_areas[['id','district']], on='id', how='left').drop('id', axis=1)
agg_public_healthindicator = agg_public_healthindicator.groupby('district').agg('mean').reset_index()

#### Save Datasets with District and Areas Data

clean_bike_trips.to_csv('../../data/processed/clean_bike_trips.csv', index=False)
clean_train_ridership.to_csv('../../data/processed/clean_train_ridership.csv', index=False)
clean_police_districts.to_csv('../../data/processed/clean_police_districts.csv', index=False)
agg_public_healthindicator.to_csv('../../data/processed/agg_public_healthindicator.csv', index=False)
disadvantaged_areas_within_areas.to_csv('../../data/processed/disadvantaged_areas_within_areas.csv', index=False)
clean_areas.to_csv('../../data/processed/clean_areas.csv', index=False)

def save_data():
    for df_name, df in data.items():
        df.to_csv(f'../../data/processed/{df_name}.csv', index=False)

save_data()
